+++ packages/icons/src/iconsets.config.ts
@@ -8,10 +8,12 @@ export default [
     name: 'post',
     apiUrl: urls.post,
     downloadDirectory: path.join(SOURCE_PATH, 'post'),
+    expectedSourcesPerIcon: 1,
   },
   {
     name: 'ui',
     apiUrl: urls.ui,
     downloadDirectory: path.join(SOURCE_PATH, 'ui'),
+    expectedSourcesPerIcon: 6,
   },
 ] as IconSet[];