+++ packages/icons/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @swisspost/design-system-icons
 
+## 9.0.0-next.17
+
+## 9.0.0-next.16
+
 ## 9.0.0-next.15
 
 ## 9.0.0-next.14