+++ packages/internet-header/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @swisspost/internet-header
 
+## 2.0.0-next.17
+
+### Patch Changes
+
+- Updated dependencies:
+  - @swisspost/design-system-styles@9.0.0-next.17
+
+## 2.0.0-next.16
+
+### Patch Changes
+
+- Updated dependencies:
+  - @swisspost/design-system-styles@9.0.0-next.16
+
 ## 2.0.0-next.15
 
 ### Patch Changes